Add files via upload
𝐓𝐢𝐜 𝐓𝐚𝐜 𝐓𝐨𝐞 𝐆𝐚𝐦𝐞'𝐬 𝐂𝐨𝐧𝐬𝐨𝐥𝐞 𝐈𝐦𝐩𝐥𝐞𝐦𝐞𝐧𝐭𝐚𝐭𝐢𝐨𝐧

𝐈𝐧𝐭𝐫𝐨𝐝𝐮𝐜𝐭𝐢𝐨𝐧:

In this game, two players will be played and you have one print board on the screen where from 1 to 9 numbers will be displayed or you can say it box number. Now,
you have to choose X or O for the specific box number. For example, if you have to select any number then for X or O will be shown on the print board, and turn for
next will be there. The task is to create a Java program to implement a 3×3 Tic-Tac-Toe game for two players.

𝐇𝐨𝐰 𝐭𝐨 𝐏𝐥𝐚𝐲 𝐭𝐡𝐞 𝐆𝐚𝐦𝐞:

● Both the players choose either X or O to mark their cells.
● There will be a 3×3 grid with numbers assigned to each of the 9 cells.
● The player who chose X begins to play first.
● He enters the cell number where he wishes to place X.
● Now, both O and X play alternatively until any one of the two wins.
● 𝐖𝐢𝐧𝐧𝐢𝐧𝐠 𝐜𝐫𝐢𝐭𝐞𝐫𝐢𝐚: Whenever any of the two players has fully filled one row/ column/ diagonal with his symbol (X/ O), he wins and the game ends.
● If neither of the two players wins, the game is said to have ended in a draw.